sequelize installed
