--- models/order.ts
@@ -1,17 +1,0 @@
-import { sequelize } from '../utils/database';
-import { Sequelize, DataTypes } from 'sequelize';
-
-const Order = sequelize.define('order', {
-  id: {
-    type: DataTypes.UUID,
-    primaryKey: true,
-    autoIncrement: true,
-    allowNull: false,
-  },
-  total: {
-    type: DataTypes.INTEGER,
-    allowNull: false,
-  },
-});
-
-export default Order;
