+++ package.json
@@ -2,8 +2,7 @@
   "name": "test",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
-  "type": "module",
+  "main": "dist/index.js",
   "scripts": {
     "start": "node dist/index.js",
     "compile": "tsc && node app.js",
