+++ src/database/models.ts
@@ -1,0 +1,4 @@
+import Customer from "./models/customer";
+import Order from "./models/order";
+
+export const models = [Order, Customer];
