+++ src/database/models/customer.ts
@@ -1,0 +1,31 @@
+import { DataTypes, Sequelize } from 'sequelize';
+import sequelize from '../database';
+
+const Customer = sequelize.define(
+  'Customer',
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+      allowNull: false,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    phone: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  },
+  {
+    timestamps: false,
+  },
+);
+
+export default Customer;
