+++ src/database/models/order.ts
@@ -1,0 +1,23 @@
+import sequelize from '../database';
+import { Sequelize, DataTypes } from 'sequelize';
+
+const Order = sequelize.define(
+  'Order',
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+      allowNull: false,
+    },
+    total: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+  },
+  {
+    timestamps: false,
+  },
+);
+
+export default Order;
